requirements.txt
aiogram==3.17.0
openai==1.59.0
apscheduler==3.11.0
python-dotenv==1.0.1
aiohttp==3.11.11
